ml_service/requirements.txt
fastapi
uvicorn
opencv-python
numpy
ultralytics
pydantic
requests
python-socketio
SpeechRecognition
pydub
pyaudio
tensorflow-cpu
tensorflow_hub
scipy

